# example/springboot-jaxrs-openapi-asciidoc/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: Task Manager
  description: Менеджер управления проектами и задачами
  contact:
    name: Денис Волненко
    url: https://www.volnenko.ru
    email: denis@volnenko.ru
  version: 1.0.0
servers:
- url: http://localhost:8080/
  description: Локальный сервер Task Manager
...
paths:
...
  /api/v1/task/one/{id}:
    get:
      tags:
      - TaskController
      summary: Получение задачи
      description: Получение задачи по Id
      operationId: selectOneTask
      parameters:
      - name: id
        in: path
        description: Идентификатор
        required: true
        schema:
          type: string
      responses:
        default:
          description: default response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDTO'
...
components:
  schemas:
...
    TaskDTO:
      type: object
      properties:
        name:
          type: string
          description: Название
        description:
          type: string
          description: Описание
        created:
          type: string
          description: Дата создания
          format: date-time
        updated:
          type: string
          description: Дата обновления
          format: date-time
        version:
          type: integer
          description: Версия
          format: int32
      description: Задача